Move curve point (un)marshaling to ecdh package
